Split SocketHandler off into its own file.
