+++ server/src/cvServer.py
@@ -25,108 +25,17 @@ simply change the position of the trackbar.
         Drawing on work done by Cleo Schneider and Audrey St. John.
 """
 
-
-import socket   # the only module needed for all our socket interactions
 import sys      # so that we can exit()
-import os       # get resource locations
 import argparse # to parse command-line arguments
 import cv2      # openCV
 import numpy    # needed for openCV
 import math     # for utility functions
 
 from utility import *
+from socket_handler import SocketHandler
 
 
-class SocketHandler():
 
-    """
-    Manages a non-blocking server socket: searches for incoming connections (checkIncomingConnections)
-    and sends messages to all found connections (sendInformation).
-    """
-
-    server = None
-    clients = []
-    message = None
-    host = None
-    port = 0
-
-    def __init__(self, host='', port=5204):
-        """
-        Creates a server socket with the specified host and port number.
-        Sets the socket to non-blocking, and sets it to listen for connections.
-        @param host: The host address. Default value: ''
-                     (gives the socket local address)
-        @param port: The port number. Default value: 5204
-                     (the port number of the majority of python applications)
-        @return: none
-        """
-
-        self.port = port
-        self.host = host
-        # Create the socket.
-        self.server = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-
-        try:
-            # Attempt to bind the socket to the host and port
-            # (Until this is done, one cannot do anything with the socket.)
-            self.server.bind((self.host, self.port))
-        except socket.error, message:
-            # If the binding failed, exit.
-            # (If the binding fails,  don't panic--if you run the code and it fails,
-            # it is quite possible that if you wait a few moments and run it again,
-            # it will work. A failure does not necessarily mean that something is
-            # wrong with the host and the port number.)
-            sys.exit()
-
-        # Sets the server to non blocking. The server needs to check for connections
-        # frequently so that all incoming requests are found, but it is quite possible
-        # that there will be no requests; in this instance, we want to continue executing
-        # instead of waiting for a connection (this is the default setting).
-        #
-        # Set the socket to listen for up to 10 incoming connections.
-        self.server.setblocking(0)
-        self.server.listen(10)
-
-    def checkIncomingConnections(self):
-        """
-        Accepts a waiting socket connection and add it to the list of clients
-        to send information to. If there is no waiting connection, do nothing.
-        (This is a non-blocking call.)
-        @return: none
-        """
-
-        try:
-            # Accept any waiting socket connections and add them to the client list.
-            client, address = self.server.accept()
-            if client is not None:
-                self.clients.append(client)
-        except socket.error, error:
-            # If there were no waiting connections, do nothing.
-            return
-
-    def sendInformation(self, message):
-        """
-        Sends the specified message to every socket connection (every item of self.clients).
-        If a connection has exited, catch the error and remove it from the list.
-        @param message: The message to send to each waiting client.
-        @return: none
-        """
-
-        toRemove = []
-        # Send the message to each client; add each exited client to toRemove.
-        for client in self.clients:
-            try:
-                client.sendall(message)
-            except socket.error:
-                # A socket.error will be thrown when we try to send a message to
-                # a socket connection that has exited. Add the connection to
-                # our list to remove after we're done iterating through the
-                # list of clients.
-                toRemove.append(client)
-
-        # Remove each item in toRemove from self.clients
-        for remove in toRemove:
-            self.clients.remove(remove)
 
 
 class AreaOfInterest:
@@ -989,7 +898,7 @@ class Data:
 
         # Add an initial default area of interest that encompasses the entire screen.
         # Although we set them to 240 by 180, they turn out to be 320 by 240 on Windows,
-	# but 240 by 180 on OS X.
+		# but 240 by 180 on OS X.
         self.addAreaOfInterest(3, 237, 3, 173)
 
         # The method set is used to set different properties of the VideoCapture instance.
@@ -1020,7 +929,7 @@ class Server:
     sWindow = "start"
     hWindow = "help window"
     helpImage = "startupHelp.png"
-    server = None
+    server_socket = None
 
 
     def __init__(self, autostart=False, cameraindex=0):
@@ -1252,7 +1161,7 @@ class Server:
         self.cameraIndex = cameraIndex
         self.data = Data(self.cameraIndex)
         # Initialize the SocketHandler, which will create the server socket.
-        self.server = SocketHandler()
+        self.server_socket = SocketHandler()
         # Create the GUI to display video feed, motion, and settings.
         self.createGUI()
         # Start the continuous loop that will run the rest of the program.
@@ -1333,8 +1242,8 @@ class Server:
             # this call would block until it found an incoming connection, but the SocketHandler
             # sets the socket to be non-blocking.
             # Once any new clients have been accepted, send all of the clients the message string.
-            self.server.checkIncomingConnections()
-            self.server.sendInformation(message)
+            self.server_socket.checkIncomingConnections()
+            self.server_socket.sendInformation(message)
 
 
 if __name__ == "__main__":
